refactor(frontend): split google account modal into setup/connected/load-state sections
501-line modal becomes a 263-line container; response schema and
external-open helpers join a google-account-model per the sibling
convention.
