+++ frontend/src/features/integrations/google-account-modal.tsx
@@ -8,256 +8,18 @@ import {
   type GoogleAccountView,
 } from "@local-studio/agent-runtime/google-account-contract";
 import type { GoogleWorkspacePluginId } from "@local-studio/agent-runtime/google-workspace-binding";
-import { Alert, Button, FormField, Input, Spinner, StatusPill, UiModal, UiModalHeader } from "@/ui";
-import { ExternalLink, KeyRound, X } from "@/ui/icon-registry";
+import { Alert, UiModal, UiModalHeader } from "@/ui";
+import { KeyRound, X } from "@/ui/icon-registry";
 import { useMountSubscription } from "@/hooks/use-mount-subscription";
-
-const GoogleCancellationResponseSchema = Schema.Struct({ cancelled: Schema.Literal(true) });
-
-function responseError(body: unknown, fallback: string): string {
-  if (!body || typeof body !== "object") return fallback;
-  const error = Reflect.get(body, "error");
-  return typeof error === "string" ? error : fallback;
-}
-
-async function requestJson<T>(
-  url: string,
-  decode: (input: unknown) => T,
-  init?: RequestInit,
-): Promise<T> {
-  const response = await fetch(url, init);
-  const body: unknown = await response.json().catch(() => null);
-  if (!response.ok) throw new Error(responseError(body, `Request failed (${response.status})`));
-  return decode(body);
-}
-
-async function openExternal(url: string): Promise<void> {
-  const bridge = window.localStudioDesktop?.openExternal;
-  if (bridge && (await bridge(url))) return;
-  if (!window.open(url, "_blank", "noopener,noreferrer")) {
-    throw new Error("Local Studio could not open the Google sign-in page");
-  }
-}
-
-function GoogleAccountLoadState({ error, onRetry }: { error: string; onRetry: () => void }) {
-  if (error) {
-    return (
-      <Alert variant="error">
-        <div className="flex flex-wrap items-center justify-between gap-3">
-          <span>{error}</span>
-          <Button variant="secondary" size="sm" onClick={onRetry}>
-            Retry
-          </Button>
-        </div>
-      </Alert>
-    );
-  }
-  return (
-    <div
-      role="status"
-      aria-live="polite"
-      className="flex min-h-24 items-center justify-center gap-2 text-sm text-(--ui-muted)"
-    >
-      <Spinner size="sm" />
-      Loading Google account
-    </div>
-  );
-}
-
-function ConnectedGoogleAccount({
-  email,
-  displayName,
-  confirming,
-  busy,
-  onConfirm,
-  onKeep,
-  onDisconnect,
-  onClose,
-}: {
-  email: string | null;
-  displayName: string;
-  confirming: boolean;
-  busy: boolean;
-  onConfirm: () => void;
-  onKeep: () => void;
-  onDisconnect: () => void;
-  onClose: () => void;
-}) {
-  return (
-    <div className="space-y-4">
-      <div
-        role="status"
-        aria-live="polite"
-        className="flex items-center justify-between rounded-lg border border-(--ui-border) px-4 py-3"
-      >
-        <div>
-          <div className="text-sm font-medium text-(--ui-fg)">{email}</div>
-          <div className="mt-1 text-xs text-(--ui-muted)">Read-only · {displayName}</div>
-        </div>
-        <StatusPill tone="good">Connected</StatusPill>
-      </div>
-      {confirming ? (
-        <>
-          <Alert variant="warning">
-            Revoking access removes every Google OAuth scope granted to this Cloud project and
-            disconnects both Gmail and Calendar. A dedicated project keeps other Google clients
-            isolated.
-          </Alert>
-          <div className="flex items-center justify-end gap-2">
-            <Button variant="secondary" onClick={onKeep} disabled={busy}>
-              Keep connected
-            </Button>
-            <Button variant="danger" onClick={onDisconnect} loading={busy}>
-              Revoke access
-            </Button>
-          </div>
-        </>
-      ) : (
-        <div className="flex items-center justify-between gap-3">
-          <Button variant="danger" onClick={onConfirm}>
-            Disconnect Google
-          </Button>
-          <Button onClick={onClose}>Done</Button>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function GoogleAccountSetup({
-  account,
-  editing,
-  clientId,
-  clientSecret,
+import {
+  GoogleCancellationResponseSchema,
+  requestJson,
+  openExternal,
   sharedClientWarning,
-  awaiting,
-  busy,
-  onClientId,
-  onClientSecret,
-  onEdit,
-  onClose,
-  onCancelSignIn,
-  onConnect,
-}: {
-  account: GoogleAccountView;
-  editing: boolean;
-  clientId: string;
-  clientSecret: string;
-  sharedClientWarning: string | null;
-  awaiting: boolean;
-  busy: boolean;
-  onClientId: (value: string) => void;
-  onClientSecret: (value: string) => void;
-  onEdit: () => void;
-  onClose: () => void;
-  onCancelSignIn: () => void;
-  onConnect: () => void;
-}) {
-  const needsClient = !account.configured || editing;
-  return (
-    <div className="space-y-4">
-      {needsClient ? (
-        <>
-          <FormField
-            label="OAuth client ID"
-            required
-            description="Use a Google Desktop OAuth client with the Workspace MCP APIs enabled."
-          >
-            <Input
-              value={clientId}
-              onChange={(event) => onClientId(event.target.value)}
-              placeholder="…apps.googleusercontent.com"
-              autoComplete="off"
-              spellCheck={false}
-            />
-          </FormField>
-          <FormField label="OAuth client secret" description="Optional for some desktop clients.">
-            <Input
-              type="password"
-              value={clientSecret}
-              onChange={(event) => onClientSecret(event.target.value)}
-              placeholder={account.hasClientSecret ? "Stored securely" : "Client secret"}
-              autoComplete="off"
-              spellCheck={false}
-            />
-          </FormField>
-          {sharedClientWarning ? <Alert variant="warning">{sharedClientWarning}</Alert> : null}
-        </>
-      ) : (
-        <div className="flex items-center justify-between rounded-lg border border-(--ui-border) px-4 py-3">
-          <div className="min-w-0">
-            <div className="text-sm font-medium text-(--ui-fg)">OAuth client ready</div>
-            <div className="mt-1 truncate text-xs text-(--ui-muted)">{account.clientId}</div>
-          </div>
-          <Button variant="ghost" size="sm" onClick={onEdit}>
-            Change
-          </Button>
-        </div>
-      )}
-      {awaiting ? (
-        <Alert variant="success">
-          Finish consent in your browser. Local Studio is checking for the connection.
-        </Alert>
-      ) : null}
-      <div className="flex flex-wrap items-center justify-between gap-3">
-        <div className="flex flex-wrap gap-1">
-          <Button
-            variant="ghost"
-            icon={<ExternalLink className="h-4 w-4" />}
-            onClick={() =>
-              void openExternal(
-                "https://developers.google.com/workspace/guides/configure-mcp-servers",
-              )
-            }
-          >
-            Setup guide
-          </Button>
-          <Button
-            variant="ghost"
-            icon={<ExternalLink className="h-4 w-4" />}
-            onClick={() => void openExternal("https://console.cloud.google.com/auth/clients")}
-          >
-            Google Cloud
-          </Button>
-        </div>
-        <div className="flex gap-2">
-          <Button
-            variant="secondary"
-            onClick={awaiting ? onCancelSignIn : onClose}
-            loading={awaiting && busy}
-            disabled={busy && !awaiting}
-          >
-            {awaiting ? "Cancel sign-in" : "Cancel"}
-          </Button>
-          <Button
-            onClick={onConnect}
-            loading={busy && !awaiting}
-            disabled={awaiting || (needsClient && !clientId.trim())}
-          >
-            {awaiting
-              ? "Waiting for Google"
-              : sharedClientWarning
-                ? "Revoke & replace"
-                : "Continue with Google"}
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function sharedClientWarning(
-  accountId: GoogleWorkspacePluginId,
-  account: GoogleAccountView | null,
-  editing: boolean,
-  clientId: string,
-): string | null {
-  const otherAccountId = accountId === "gmail" ? "google-calendar" : "gmail";
-  if (!editing || !account?.connections[otherAccountId].connected) return null;
-  if (clientId.trim() === account.clientId) return null;
-  const otherDisplayName = accountId === "gmail" ? "Google Calendar" : "Gmail";
-  return `Replacing this client revokes the current Cloud project's Google access and disconnects ${otherDisplayName} before starting again.`;
-}
+} from "./google-account-model";
+import { GoogleAccountLoadState } from "./google-account-load-state";
+import { ConnectedGoogleAccount } from "./google-account-connected";
+import { GoogleAccountSetup } from "./google-account-setup";
 
 export function GoogleAccountModal({
   accountId,
